Move logic into deephaven_internal.jvm
