+++ py/embedded-server/deephaven_server/start_jvm.py
@@ -7,7 +7,7 @@ import itertools
 import os
 import pathlib
 
-import deephaven_internal
+from deephaven_internal import jvm
 
 # TODO(deephaven-core#2592): Generalize start_jvm to work with importlib.resources
 
@@ -91,7 +91,7 @@ def start_jvm(
     else:
         jvm_args.extend(required_jvm_args)
 
-    deephaven_internal.init_jvm(
+    jvm.init_jvm(
         java_home=java_home,
         # jvm_dll=jvm_dll,
         jvm_classpath=jvm_classpath,
